Refactor(#21): 파라미터 id값 타입 int -> long 변경

+++ application/wypl-core/src/main/java/com/wypl/wyplcore/review/data/response/ReviewIdResponse.java
@@ -6,7 +6,7 @@ public record ReviewIdResponse(
 	@JsonProperty("review_id")
 	long reviewId
 ) {
-	public static ReviewIdResponse from(int reviewId) {
+	public static ReviewIdResponse from(long reviewId) {
 		return new ReviewIdResponse(reviewId);
 	}
 }
